moving files around and some changes for deployment
