+++ BudgetAutomation/BudgetAutomation.Engine/LocalTesting/Program.Internals.cs
@@ -1,4 +1,4 @@
-﻿namespace BudgetAutomation.Engine.Misc;
+﻿namespace BudgetAutomation.Engine.LocalTesting;
 
 // Partial class used for UnitTests to have access to Program
 public partial class Program { }
