+++ BudgetAutomation/BudgetAutomation.Engine/Program.cs
@@ -1,4 +1,5 @@
 using BudgetAutomation.Engine.Extension;
+using BudgetAutomation.Engine.LocalTesting;
 using BudgetAutomation.Engine.Misc;
 using Microsoft.Extensions.Options;
 using SharedLibrary.Settings;
