+++ BudgetAutomation/TelegramListener/Startup.cs
@@ -1,5 +1,3 @@
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
 using TelegramListener.Extension;
 
 namespace TelegramListener;
@@ -10,7 +8,7 @@ public class Startup
     public void ConfigureServices(IServiceCollection services)
     {
         // Local development settings
-        var isLocalDev = SharedLibrary.LocalDevelopment.SamStart.IsLocalDev();
+        var isLocalDev = false;//SharedLibrary.LocalTesting.SamStart.IsLocalDev();
 
         var config = new ConfigurationBuilder().AddProjectSpecificConfigurations(isLocalDev).Build();
