+++ BudgetAutomation/UserManagerApi/Startup.cs
@@ -1,27 +1,10 @@
-using Amazon;
-using Amazon.DynamoDBv2;
-using Amazon.DynamoDBv2.DataModel;
-using Amazon.Lambda.Serialization.SystemTextJson;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using SharedLibrary.Settings;
-using UserManagerApi.AotTypes;
 using UserManagerApi.Extension;
-using UserManagerApi.Service;
 
 namespace UserManagerApi;
 
 [Amazon.Lambda.Annotations.LambdaStartup]
 public class Startup
 {
-    /// <summary>
-    /// Services for Lambda functions can be registered in the services dependency injection container in this method. 
-    ///
-    /// The services can be injected into the Lambda function through the containing type's constructor or as a
-    /// parameter in the Lambda function using the FromService attribute. Services injected for the constructor have
-    /// the lifetime of the Lambda compute container. Services injected as parameters are created within the scope
-    /// of the function invocation.
-    /// </summary>
     public void ConfigureServices(IServiceCollection services)
     {
         services.AddProjectSpecificServices(new ConfigurationManager());
